Implemented hasball and a very crude dribbler for robot. Made the ball have lesslinear damping and decreased it's mass a little bit.
Dribbling is  very complex. checked in code simply applies a torque to the ball if the ball is located close engough and right in front of the robot. Problem isthat if robot/ball friction is high, the dribbling tends to bounce the ball. If robot/ball friction is low, robot has ball rotate sideways and away with the slightest robot rotation while dribbling. Seems non-ideal. Likely will need to reconstruct the actual physical dribbling device, with a spinning roller. Is possible that the behaviour here is acutally not too far off the real behaviour.

-Jason